Refactor RestTest class to separate OperatiorRestTest and WebhookRestTest casses
